+++ examples/advanced/fedot_profiler_example.py
@@ -4,7 +4,7 @@ import random
 import numpy as np
 
 from cases.credit_scoring.credit_scoring_problem import run_credit_scoring_problem, get_scoring_data
-from fedot.utilities.profiler.fedot_memory_profiler import MemoryProfiler
+from fedot.utilities.profiler.automl_memory_profiler import MemoryProfiler
 from fedot.utilities.profiler.time_profiler import TimeProfiler
 
 random.seed(1)
