+++ fedot/utilities/profiler/automl_memory_profiler.py
