+++ test/integration/test_profiler.py
@@ -5,7 +5,7 @@ import shutil
 import pytest
 
 from cases.credit_scoring.credit_scoring_problem import get_scoring_data, run_credit_scoring_problem
-from fedot.utilities.profiler.fedot_memory_profiler import MemoryProfiler
+from fedot.utilities.profiler.automl_memory_profiler import MemoryProfiler
 from fedot.utilities.profiler.time_profiler import TimeProfiler
